Move example to static
remove nrlmsis related stuff from dockerfile
put examples in the right directory from start
