+++ src/odinapi/static/examples/get_l2_data.py
